Moving logic out of components
